requirements.txt
python-evtx
